+++ src/com/Data/Graph.java
@@ -5,7 +5,7 @@ import java.util.Map;
 public class Graph {
     private Map<Integer, Map<Integer, RoadInfo>> edges = new HashMap<>();
 
-    public void addEdges(int src, int dest, int distance, int maxVelocity) {
+    public void addEdges(int src, int dest, int distance, double maxVelocity) {
             RoadInfo road = new RoadInfo(distance, maxVelocity, src, dest);
             if (edges.get(src) == null) {
                 HashMap<Integer, RoadInfo> adjacents = new HashMap<>();
